refactor(project): refactor project structure
