+++ src/app/(insert-ad-page)/weather/components/client/WeatherImage.tsx
