+++ src/app/(insert-ad-page)/weather/components/client/bottomSection/RecentPlace.tsx
