+++ src/app/(insert-ad-page)/weather/components/client/bottomSection/WeatherAnimation.tsx
