+++ src/app/(insert-ad-page)/weather/components/client/bottomSection/index.tsx
