+++ src/app/(insert-ad-page)/weather/components/client/bottomSection/searchbar/index.tsx
