+++ src/app/(insert-ad-page)/weather/components/client/index.ts
