+++ src/app/(insert-ad-page)/weather/components/server/ThirtySixHoursWeather.tsx
@@ -2,7 +2,7 @@ import errorHandler from "@/utils/errorHandler"
 import { getWeatherApiEndpoint } from "../../utils/getWeatherApiEndpoint"
 import { styled } from "@pigment-css/react"
 import { WeatherAPIResponse } from "../../types"
-import Humidity from "../../../../assets/weather/humidity.svg"
+import Humidity from "../../../../../assets/weather/humidity.svg"
 import createWeatherCardList from "../../utils/createWeatherCardList"
 import { WeatherCardTime, WeatherImage } from "../client"
 import { getWeatherType } from "../../utils/getWeatherType"
