+++ src/app/(insert-ad-page)/weather/constant.ts
