+++ src/app/(insert-ad-page)/weather/types.ts
