+++ src/app/(insert-ad-page)/weather/utils/checkIsAvailableCity.ts
