+++ src/app/(insert-ad-page)/weather/utils/createWeatherCardList.ts
