+++ src/app/(insert-ad-page)/weather/utils/getCityCardList.ts
