+++ src/app/(insert-ad-page)/weather/utils/getIsDayOrNight.ts
