+++ src/app/(insert-ad-page)/weather/utils/getWeatherApiEndpoint.ts
