+++ src/app/(insert-ad-page)/weather/utils/getWeatherType.ts
