+++ src/app/layout.tsx
@@ -3,7 +3,6 @@ import { Inter } from "next/font/google"
 import "@pigment-css/react/styles.css"
 import "./globals.css"
 import NavBar from "./NavBar"
-import { styled } from "@pigment-css/react"
 
 const inter = Inter({ subsets: ["latin"] })
 
@@ -11,17 +10,6 @@ export const metadata: Metadata = {
 	title: "Create Next App",
 	description: "Generated by create next app",
 }
-
-const ContentWrapper = styled("div")({
-	width: "70%",
-	height: "calc(100vh - 70px)",
-	border: "1px solid black",
-	margin: "0 auto",
-	backgroundColor: "#FFF",
-	"@media (max-width: 1024px)": {
-		width: "100%",
-	},
-})
 
 export default function RootLayout({
 	children,
@@ -37,7 +25,7 @@ export default function RootLayout({
 				}}
 			>
 				<NavBar />
-				<ContentWrapper>{children}</ContentWrapper>
+				{children}
 			</body>
 		</html>
 	)
